backend/ML/requirements.txt
h5py
opencv-python
numpy
matplotlib
keras
tensorflow-gpu